getProfileReports moved to profileController
