+++ controllers/reports-controller.js
@@ -82,23 +82,6 @@ const updateReport = (req, res) => {
 		});
 };
 
-const getProfileReports = (req, res) => {
-	knex("reports")
-		.where({ user_id: req.params.id })
-		.then((reports) => {
-			if (reports.length === 0) {
-				res
-					.status(404)
-					.json({ message: `no reports with user ID: ${req.params.id} exist` });
-			} else {
-				res.send(reports);
-			}
-		})
-		.catch((err) => {
-			res.status(500).json({ message: `${err}` });
-		});
-};
-
 const getReportID = (req, res) => {
 	knex("reports")
 		.where({ id: req.params.id })
